test: Suppress pom "Skipping Coordinate" msgs
During tests, when writing a pom file, "Skipping Coordinate" messages
for our `:local/root` jar files are written to stderr.

We don't care that these coords are skipped for our pom.
We are creating a test uberjar.

But each time I come back to this project, being forgetful, I see these
warnings and wonder if I should care.

So, as a favour to future us, suppress them.